Consolidate EVM asset logic
