--- wallet.js
@@ -1,52 +1,0 @@
-import { escapeHtml } from './lib.js';
-
-export function formatConnectedTokenAmount(value) {
-  const amount = Number(value);
-  if (!Number.isFinite(amount)) return String(value ?? '0');
-  if (amount === 0) return '0';
-  if (Math.abs(amount) < 0.000001) {
-    return amount.toExponential(4);
-  }
-  return amount.toLocaleString(undefined, {
-    maximumFractionDigits: 6,
-    minimumFractionDigits: 0,
-  });
-}
-
-export function formatConnectedUsd(value) {
-  if (value === null || value === undefined || value === '') {
-    return 'Value unavailable';
-  }
-  const amount = Number(value);
-  if (!Number.isFinite(amount)) return 'Value unavailable';
-  return amount.toLocaleString(undefined, {
-    style: 'currency',
-    currency: 'USD',
-    maximumFractionDigits: amount > 0 && amount < 0.01 ? 6 : 2,
-  });
-}
-
-export function connectedAssetLogoMarkup(asset, walletNetwork) {
-  const logoUrl = typeof asset.logoUrl === 'string' ? asset.logoUrl : '';
-  if (/^(?:https:\/\/|\.\/)/.test(logoUrl)) {
-    return `<img src="${escapeHtml(logoUrl)}" alt="" class="connected-asset-logo-image">`;
-  }
-  return `<span class="connected-asset-logo-fallback">${escapeHtml(walletNetwork.shortName.slice(0, 3))}</span>`;
-}
-
-export function formatAssetDetailsUpdatedAt(timestamp) {
-  if (!timestamp) return 'Updated just now';
-  return `Updated ${new Intl.DateTimeFormat(undefined, {
-    month: 'short',
-    day: 'numeric',
-    hour: 'numeric',
-    minute: '2-digit',
-  }).format(new Date(timestamp))}`;
-}
-
-export function formatConnectedTokenType(asset) {
-  const type = String(asset?.tokenType || '').toLowerCase();
-  if (type === 'native') return 'Native asset';
-  if (type === 'erc20') return 'ERC-20';
-  return type ? type.toUpperCase() : 'Token';
-}
